requirements.txt
Flask==3.0.0
Pillow
qrcode==7.4.2
python-dotenv==1.0.0
Werkzeug==3.0.0
Flask-WTF==1.2.1
gunicorn==21.2.0